requirements.txt
Click==7.0
Flask==1.1.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
mock==4.0.1
PyYAML==5.3
SQLAlchemy==1.3.13
Werkzeug==1.0.0
WTForms==2.2.1
yml==0.0.1
